Initial commit; unit tests required.
